candidate_data_cleaning/pyproject.toml
[project]
name = "candidate-data-cleaning"
version = "0.1.0"
description = "Reusable pipeline to clean candidate datasets for the headhunter collaboration platform"
authors = [{ name = "Lillian Liao" }]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
